chore: componentize run inspector panel
